+++ modules/core/artfish_overview_explorer_server.R
@@ -47,7 +47,7 @@ artfish_overview_explorer_server <- function(id, parent.session, lang = NULL, po
         ref_species = artfish$ref_species()
       )
     })
-
+    
     artfish_shiny_overview_server(
       "artfish_overview_explorer",
       lang = lang,
@@ -55,7 +55,8 @@ artfish_overview_explorer_server <- function(id, parent.session, lang = NULL, po
       effort_source = artfish$effort_source,
       minor_strata = artfish$minor_strata,
       opts = list(
-        refresh_ui = actionButton(ns("refresh_artfish_estimates"), icon = icon("refresh"), label = "")
+        refresh_ui = actionButton(ns("refresh_artfish_estimates"), icon = icon("refresh"), label = ""),
+        values_ui = if(sum(estimate_r()$trade_value, na.rm = T) == 0) FALSE else TRUE
       )
     )