+++ modules/core/artfish_species_server.R
@@ -22,7 +22,11 @@ artfish_species_server <- function(id, parent.session, lang = NULL, pool, reload
   INFO("Get Artfish computation outputs for UI")
   estimate <- get_artfish_results_for_ui(input=files,input_type = "file", ref_fishing_units, ref_species)    
   
-  artfishr::artfish_shiny_species_server("artfish_species", lang = lang, estimate = reactive({ estimate }))
+  artfishr::artfish_shiny_species_server("artfish_species", lang = lang, 
+                                         estimate = reactive({ estimate }),
+                                         opts = list(
+                                           values_ui = if(sum(estimate$trade_value, na.rm = T) == 0) FALSE else TRUE
+                                         ))
   
   MODULE_END_TIME <- Sys.time()
   INFO("artfish-species: END")